+++ crates/egui_router/examples/async_router.rs
@@ -21,8 +21,12 @@ async fn main() -> eframe::Result<()> {
         "Router Example",
         NativeOptions::default(),
         move |ctx, _frame| {
+            ctx.all_styles_mut(|style| {
+                style.interaction.selectable_labels = false;
+            });
             let router = router.get_or_insert_with(|| {
                 EguiRouter::builder()
+                    .swipe_back_gesture(true)
                     .error_ui(|ui, state: &AppState, error| {
                         ui.label(format!("Error: {error}"));
                         if ui.button("back").clicked() {