+++ crates/egui_router/src/lib.rs
@@ -40,7 +40,7 @@ impl<F: FnMut(&mut Ui, &mut State), State> Route<State> for F {
 static ID: AtomicUsize = AtomicUsize::new(0);
 
 struct RouteState<State> {
-    path: String,
+    path_with_query: String,
     route: HandlerResult<Box<dyn Route<State>>>,
     id: usize,
     state: u32,
@@ -96,12 +96,12 @@ impl TransitionConfig {
         }
     }
 
-    /// A iOS-like slide transition (Same as [`TransitionConfig::default`])
+    /// An iOS-like slide transition (Same as [`TransitionConfig::default`])
     pub fn slide() -> Self {
         Self::default()
     }
 
-    /// A android-like fade up transition
+    /// An Android-like fade up transition
     pub fn fade_up() -> Self {
         Self::new(
             SlideFadeTransition(